test: refactor test infrastructure and modernize database tests
Extract external dependency mocks into a dedicated `tests/mocks.py` module to clean up `conftest.py`. Introduce `ble_cleanup.py` and `sqlite_provenance.py` to improve test teardown and detect leaked SQLite connections.

Additionally, migrate `test_db_runtime_security.py` from `unittest` to `pytest` fixtures and remove the redundant `test_e2ee_integration.py` script.
